[bindings] merge dunext into dunegdt
